Time: 27 ms (5.82%) | Memory: 45.7 MB (93.63%) - LeetSync

+++ 9-palindrome-number/palindrome-number.java
@@ -1,7 +1,7 @@
 class Solution {
 
      public static long reversenum(int x){
-        long rev = 0;
+        int rev = 0;
 
         while(x != 0){
             rev = rev * 10 + (x % 10);
